Implement personality function for Runa exceptions
As a first pass, this is implemented with C code cobbled together mostly from
the GCC project. I spent a few years (not full-time, happily...) trying to get
it working with Runa code, but that has proved to be extremely finicky. Noting
that even the Rust project calls out to C for this stuff (after specifically
mentioning how tricky it is), I'm satisfied to get something working for now!

For now, exception filtering is one prominent example of something that
doesn't work yet; this will be fixed in the future, I guess.